HV:misc:fix "signed/unsigned conversion without cast"
Misra C required signed/unsigned conversion with cast.

V1->V2:
  a.split patch to patch series

V2->V3:
  a.change the uint64_t type numeric constant's suffix from U to UL

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/gdt.c
@@ -7,7 +7,7 @@
 #include <hypervisor.h>
 
 static void set_tss_desc(union tss_64_descriptor *desc,
-		void *tss, int tss_limit, int type)
+		void *tss, size_t tss_limit, int type)
 {
 	uint32_t u1, u2, u3;
